add abstract AnInteger subject with Byte, Short, Integer and Long implementations

+++ core/src/test/java/org/truth0/IterationOverSubjectTest.java
@@ -17,7 +17,7 @@
 package org.truth0;
 
 import static org.truth0.Truth.ASSERT;
-import static org.truth0.subjects.IntegerSubject.INTEGER;
+import static org.truth0.subjects.LongSubject.LONG;
 import static org.truth0.subjects.StringSubject.STRING;
 
 import java.util.Arrays;
@@ -37,9 +37,9 @@ public class IterationOverSubjectTest {
 
   @Test public void collectionItemsAreGreaterThan() {
     Iterable<Long> data = iterable(2l, 5l, 9l);
-    ASSERT.in(data).thatEach(INTEGER).isNotEqualTo(4);
+    ASSERT.in(data).thatEach(LONG).isNotEqualTo(4);
     try {
-      ASSERT.in(data).thatEach(INTEGER).isNotEqualTo(9);
+      ASSERT.in(data).thatEach(LONG).isNotEqualTo(9);
       ASSERT.fail("Expected assertion to fail on element 3.");
     } catch (AssertionError e) {
       if (e.getMessage().startsWith("Expected assertion to fail")) {
